feat/#90: article list 관련 상수 정리 및 공통 컴포넌트로 분리
